requirements.txt
pytest
requests~=2.32.3
allure-python-commons~=2.13.5
selene==2.0.0rc9
allure-pytest
python-dotenv
jsonschema